Refactor: Clean Architecture improvements and ongoing review.
Completed Steps:
1. Reviewed Domain Layer:
    - No major violations. I noted a minor point on the User entity's direct use of HttpError/HttpStatusCode.
2. Reviewed Application Layer:
    - I identified and addressed a dependency violation: AuthUseCase was importing IoC TYPES from Infrastructure.
3. Refactored IoC Type Dependencies:
    - I created `src/domain/ioc.types.ts` and `src/application/ioc.types.ts`.
    - I moved relevant IoC type definitions (DOMAIN_TYPES, APPLICATION_TYPES) to their respective layers.
    - I updated `AuthUseCase` to use these new, correctly scoped types.
    - I updated `src/infraestructure/ioc/config.ts` to use the new type definitions.
    - I renamed `TYPES` in `src/infraestructure/ioc/types.ts` to `INFRASTRUCTURE_TYPES` and retained only infra-specific types.
4. Reviewed Infrastructure Layer:
    - I identified that `PostgresUserRepository` uses an outdated IoC token.
    - I identified that `IMailService` is defined but not implemented (you requested to leave this as is).
    - I noted missing standard 404 middleware.

Work in Progress / Next Steps (before your feedback to stop):
- I was about to correct the IoC token in `src/infraestructure/database/postgres/postgres.user.repository.ts`.
- I planned to implement the standard 404 middleware as you requested.
- I planned to skip any changes to `IMailService` as per your feedback.

This work includes the significant IoC refactoring. The correction for `PostgresUserRepository` and the 404 middleware implementation were planned for subsequent steps.
