+++ src/application/ioc.types.ts
@@ -1,0 +1,3 @@
+export const APPLICATION_TYPES = {
+    IJwtService: Symbol.for('IJwtService')
+};
