+++ src/infraestructure/ioc/types.ts
@@ -1,11 +1,4 @@
-export const TYPES = {
-    //* Repositories
-    IUserRepository: Symbol.for('IUserRepository'),
-
-    //* Services
-    IHashingService: Symbol.for('IHashingService'),
-    IJwtService: Symbol.for('IJwtService'),
-
-    //* DB
+export const INFRASTRUCTURE_TYPES = {
     PostgresPool: Symbol.for('PostgresPool')
-}
+    // Add other infrastructure-specific symbols here if any in the future
+};
